pyproject.toml
[project]
name = "attendance-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=20.7",
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "groq>=0.4.2",
    "pymongo>=4.6.1",
    "motor>=3.3.2",
    "pydantic>=2.5.3",
    "python-dateutil>=2.8.2",
    "python-dateutil>=2.8.2",
    "langchain-community>=0.0.10",
    "langchain==0.1.20",
    "langchain-openai>=0.0.2",
]
